Added luxon time, creared module to switch pages
